+++ CHANGELOG.md
@@ -5,7 +5,7 @@ All notable changes to this project will be documented in this file.
 The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/),
 and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [Unreleased]
+## [1.2.0] - 2022-01-15
 
 ### Added
 
@@ -39,5 +39,6 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 - First release
 
+[1.2.0]: https://github.com/lxbrvr/alfred-keepassxc-workflow/compare/1.1.0...1.2.0
 [1.1.0]: https://github.com/lxbrvr/alfred-keepassxc-workflow/compare/1.0.0...1.1.0
 [1.0.0]: https://github.com/lxbrvr/alfred-keepassxc-workflow/releases/tag/1.0.0